+++ src/datumaro/plugins/data_formats/tabular.py
@@ -115,7 +115,13 @@ class TabularDataBase(DatasetBase):
                 target_dtype = table.dtype(target_)
                 if target_dtype in [int, float, pd.api.types.CategoricalDtype()]:
                     # 'int' can be categorical, but we don't know this unless user gives information.
-                    labels = set(table.features(target_, unique=True))
+                    labels = set(
+                        [
+                            feature
+                            for feature in table.features(target_, unique=True)
+                            if not pd.isna(feature)
+                        ]
+                    )
                     if category is None:
                         categories.add(target_, target_dtype, labels)
                     else:  # update labels if they are different.